requirements.txt
Flask
Flask-SocketIO
pywin32
numpy
opencv-python